Add files via upload
Added support for Ncomp nodes.
When the [levelname].qb is loaded it looks for a [levelnam].q file, if the file is found it check the first byte and if that byte is 0xAA it will try to parse the .q file. The .q file is terminated when 0x0 is reached.

Current bugs:
After loading a level with ncomp nodes you will freeze game after you want to change level again.