New size limits
This is needed for new limits of scripts and nodearray. Old limit of nodearray was 4096 nodes, new limit is 16384 nodes.
The script heap limit is also increased by a ton and will probably not need to be increased anymore.
Some other limits were increased aswell, more info on that will be added later on if needed.