Fixed bugs in LM_GetParam and GetParam function
There were bugs in the function that made it not parse if a struct was within an array that was within a struct.
This fix made the settings menu work as it should.
Also I changed the text of air trick speed, so it's more clear what the option does.